+++ samples/snippets/standard/io/file-names/cs/rename.cs
@@ -1,0 +1,10 @@
+﻿using System.IO;
+
+class Example3
+{
+    static void Main()
+    {
+        var fi = new FileInfo(@".\test.txt");
+        fi.MoveTo(@".\Test.txt");
+    }
+}
