--- samples/snippets/standard/io/file-names/cs/rename/rename.cs
@@ -1,10 +1,0 @@
-﻿using System.IO;
-
-class Example
-{
-   static void Main()
-   {
-      var fi = new FileInfo(@".\test.txt");
-      fi.MoveTo(@".\Test.txt");
-   }
-}
